powerpc/powernv: Remove pnv_ioda_unset_peltv()
The function was added to support the hot-remove case where we need to tear
down the PE and any associations made between that PE and other PEs which
are above it in the heirachy. However, the the function we use to setup the
PELT-V, pnv_ioda_set_peltv(), handles the remove case too.

This should also fix a bug where some associations would remain if the
PE being torn down was a compound PE.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>