wip: use block_cfg_map in the config accessors rather than looking at the EEH state
we could probably make this a per-PE bit and stash it as a flag in
pnv_ioda_pe. The attraction of using a per-bdfn bitmap is that we can also
use it to prevent spurious eeh's on p7-ioc since we need to PE map entire
buses at once there.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>

blah