powerpc/eeh: Drop the full rescan recovery path
To recover a pci device EEH can use one of two methods: If the driver
implements the error handling callbacks then we can rely on the driver to bring
the device back to a working state. If the driver doesn't implement error
handling then the device needs to be torn down and re-probed. The main
difference between these two is that in the former case the pci_dev is
preserved while in the latter a new pci_dev is created.

Recovery operations occur on a per-PE basis rather than a per-device basis so
EEH has two recovery paths split along the lines above:

1. If there are devices in the PE where we need to preserve the pci_dev then
   walk the devices in the PE and selectively remove those which don't support
   recovery. We may have child buses in this case which we need to preserve
   so in this approach we will always keep the pci_dev of bridges.

2. If none of the devices in the PE, or in child PEs, support recovery then
   remove every pci_dev under the PE's primary bus, do the reset, and rescan.

There are also a few other misc conditions which require 1) to be used instead
of 2) even if none of the drivers support recovery. Namely, when there are
devices in the PE which are passed through to a guest and when the PE only
contains a VF. In the pass-through case the guest controls the PE state rather
then the hypervisor so we want to preserve the device state, otherwise we may
cause issues in the guest. Similarly, in the VF case the VF may share a bus with
it's parent PF so we can't simply remove all the devices on that bus.

All things considered it's not entirely clear why we want or need 2). The
edge cases mentioned above mean that 2) can only be used a limited set of
circumstances. Practically speaking, supporting 2) does nothing but add
complexity to the already complicated implementation of EEH, so why keep
it?

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>