powerpc/eeh: Drop the full rescan recovery path
EEH has two methods for recovering PCI devices. When the device's driver
implements the error handling callbacks we assume the driver can handle
restoring the device to a working state. When there's no driver support EEH
handles recoverying the device by tearing down the device, resetting it,
and then re-scanning. That way the PCI core handles re-initialising the
device during a rescan ensures the devices comes back up in a working
state.  The main difference between the two methods that in the second we
will have different pci_dev structures before and after recovery, while in
the first we always have the same pci_dev.

When recovering an EEH PE we need to decide which method to use for each
device in that PE. This requires traversing all the devices which would be
recovered by the EEH pass (all the devices in the PE and any child PEs) and
checking if each device driver supports error recovery. If there are no
devices in the PE that support recovery we can use the normal PCI hotplug
infrastructure to recursively remove and rescan all the devices under that
PE. However, if there any devices that support recovery we need to be a
little more careful.

The main issue is that we need to preserve the PCI bus structure inside the
PE. This is required since a pci_dev to be recovered may be on a leaf bus.
In that case we need to ensure that all the bridges (and buses) leading to
the device are also preserved. This is generally handled by the platform
since most PCI-PCI bridges won't have a driver, let alone one that supports
the error handling callbacks. This works largely because the EEH core will
preserve and restore the first 64 bytes of config space, which covers the
bridge window registers. There is similar special handling required for
passed-through devices too. For passed through devices the guest is
responsible for managing the state of the PE so we need to leave any
hypervisor state intact (i.e the pci_devs and bus structure).

The net result is that we have two, completely independent, paths for
implementing EEH PE recovery. The full-rescan path is only really usable in
a limited set of circumstances. Although it is simpler than the
partial-rescan case that simplicity doesn't really get us anything. The
core feature of EEH is allowing device drivers to recover from errors.  The
partial-rescan case is used to implement that feature so that is one we
actually need to support.

The full-rescan case also ends up causing some other issues due to its
recursive nature. If a device is manually removed by the user (e.g. via
sysfs) the full rescan results in the device popping back into existence.
This was first noticed in d2b0f6f77ee5 ("powerpc/eeh: No hotplug on
permanently removed dev") which added the EEH_DEV_REMOVED flag to eeh_dev
in order to avoid re-scanning devices that were removed after exceeding
the max failures threshold. It is still an issue for devices removed by
other means.

To fix this drop the full re-scan path. This leave us with one code path
used by all EEH operations. Having one code path should simplify the logic
of EEH recover process and allow us to better define how the EEH core
is supposed to interact with the platform EEH backend and the PCI core.

Eliminating the recursive remove / rescan also allows us to limit the scope
of a recovery. This will be useful for implementing parallel recovery of
child PEs (e.g. VFs).

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>