powerpc/eeh: Use a goto for recovery failures
Currently eeh_handle_normal_event() has some fairly strange control flow. When
a error occurs during recovery we mark all the devices in a PE as permernantly
failed, but we do so in a fairly indirect manner. In each case we set the
recovery status to PCI_ERS_RESULT_DISCONNECTED, then that eventually falls
through into the error case handler.

It's not the most byzantine part of EEH, but it does make it difficult to
differentiate between the good and bad paths. Drop the (ab)use of the result
field and use a goto to jump straight to where we handle the failure before
exiting.

This also lets us de-indent^Wleft-shift the function a bit.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>