powerpc/eeh: Remove redundant VF PE handling
The VF PE handling in eeh_reset_devices() is redundant since the final phase of
eeh_handle_normal_event() will walk the rmv_data.removed_vf_list and re-add
each. In theory that code handles the slightly different case where a PE that
happens to contains VFs in child PEs is being recovered, but in practice it
handles both situations.

The special case in eeh_reset_devices() isn't even complete since it fails
to take the VF that it adds off the removed VF list and relies on the other
code to do that. Might as well get rid of it and handles VFs in the one
place.

For the record: We have seperate VF handling since if the PF is reset then
the PF's driver needs to re-enable SR-IOV. This seems to happen in the
.slot_reset() callback so we re-add any removed VFs after that's done, but
before calling .resume() for the PF driver. In the case where the removed
VF is in a VF PE (i.e. errors are isolated to just that VF) we don't have
that problem, but doing it there doesn't hurt either.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>