powerpc/eeh: Remove redundant VF PE handling
The VF PE handling in eeh_reset_devices() is redundant since the final
phase of eeh_handle_normal_event() will walk the rmv_data.removed_vf_list
and re-add each. The intention seems to have been that eeh_reset_devices()
would handle the special case of VF PEs which contain only a single VF
device. However, the eeh_handle_normal_event() handles that case just fine.

It's worth worth pointing out that the VF handling in eeh_reset_devices()
is broken. After re-adding the VF pci_dev, it does not delete the VF's
eeh_dev from the list of removed VFs so when we return to
eeh_handle_normal_event() the VF device will be added again. This doesn't
cause any problems since the SR-IOV core checks for an existing pci_dev
before creating a new one for the VF. It does however highlight the
pointlessness of the VF handling in eeh_reset_devices().

For the record: We have seperate VF handling since if the PF is reset then
the PF's driver may need to re-enable SR-IOV. This can be done in either
the PF's .slot_reset() or .resume() callback, so we need to ensure that is
called before attempting to re-add any removed VF devices.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>