powerpc/eeh: Rename eeh_enabled() to eeh_detection_enabled()
The various EEH detection hooks are the only remaining users of
eeh_enabled(). In those it's used to short-circuit the EEH checks in the
MMIO and config space accessor functions when there are no EEH capable
devices in the system. On PowerNV its also used to check if we need to
have the OPAL EEH event "interrupt" enabled.

Rename the function to better reflect what it's actually used for.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>