powerpc/eeh: Do not gate sysfs files on eeh_enabled()
In the distant past adding the sysfs files was done using an initcall. This
ran after the EEH platform was initialized and PCI devices created. In that
context it made sense to check the EEH enabled flag since if there was no
EEH devices in the system we'd want to avoid adding the sysfs files at all.

Adding the sysfs files was later moved so it occured as a the normal
pci_dev setup/teardown when the device is on an EEH capable bus. This
means that the EEH sysfs helpers are only ever called once we've already
determined the device supports EEH so the check is pointless. Remove it.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>