powerpc/eeh: Remove spurious checks in eeh_remove_device()
eeh_remove_device() is only called via a pci_bus notifier when a pci_dev is
removed from its bus. In that situation the pci_dev pointer passed to the
function is guaranteed to be a valid so performing a NULL check is
pointless.

Similarly, checking eeh_enabled() is unnecessary and probably broken.
If the EEH_FORCE_DISABLE flag is set at runtime via the eeh_enabled debugfs
file then we'll skip the EEH device teardown entirely. If the EEH state for
the device was initialised before "disabling EEH" we'll leave the edev in
the EEH PE tree with edev->pdev pointing to a now free()ed object.

I'm not entirely sure how "disabling" EEH at runtime is supposed to work.
However, I'm pretty sure the current behaviour isn't what anyone wants.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>